refactor(core): migrate build process to tsdown config
