+++ core/ai-commit-msg/tsconfig.json
@@ -1,6 +1,11 @@
 {
   "extends": ["../../tsconfig.base.json"],
-  "include": ["./src/**/*", "vitest.config.ts", "build.ts", "./test/**/*"],
+  "include": [
+    "./src/**/*",
+    "vitest.config.ts",
+    "tsdown.config.ts",
+    "./test/**/*"
+  ],
   "compilerOptions": {
     "types": ["node"]
   }
