+++ core/auto-readme/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": ["../../tsconfig.base.json"],
-  "include": ["./src/**/*", "build.ts", "cli.mjs"],
+  "include": ["./src/**/*", "tsdown.config.ts", "cli.mjs"],
   "compilerOptions": {
     "types": ["./types/plugins.d.ts", "node"],
     "composite": true
