+++ core/eslint-config/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@stephansama/eslint-config",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "description": "stephansama eslint configuration for multiple use cases",
   "keywords": [
     "eslint-config"
