+++ core/multipublish/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": ["../../tsconfig.base.json"],
-  "include": ["./src/**/*", "build.ts", "vitest.config.ts"],
+  "include": ["./src/**/*", "tsdown.config.ts", "vitest.config.ts"],
   "compilerOptions": {
     "composite": true
   }
